[#2374] Move Dashboard container component into components/dashboard
